move exec config to util module
